refactor: delete zero-caller packages and unused exports (#360)
- Delete quant/, accel/, cache/, compile_helpers/, parallelism/, engines/
  (with the never-satisfied SDK-engine gating branch + orphaned
  _resolve_batched_model_path in worker.py), conversion/dtype_utils.py,
  presets.py, and the accel capability test.
- Prune top-level __init__ exports with zero consumers (batched_inference
  re-export, Clamp, prompt-role types, MediaAsset, Compute, Tensors,
  load_loras, with_oom_retry, streaming signal types, Binding, unused
  error classes); internal modules import from their real homes.
- runtimes/ar_tts: keep lookup() (chatterbox-tts, musicgen); drop
  register/all_specs and the vllm_arch-less bark entry.
- micro_batch: document the breaks_cross_request_batching attr convention
  inline now that gen_worker.cache is gone; delete the acceleration
  cookbook that taught the deleted packages.
- clone/: delete dead dedup scaffolding (maybe_noop, preflight_clone,
  MirrorPreflight, vestigial execute_miss wrapper + dedupe.skipped emit),
  the 4 NotImplementedError intermediate-API stubs and their __init__
  wrappers, and dedupe the twin _tensors_artifact helpers.

pipeline/loader.py item deliberately skipped — deferred to #366 which
deletes the module wholesale. The _finalize_clone/_finalize_publish_as_is
single-path merge is deferred to #367 (cozy_convert extraction) which
rebuilds finalize with its first test coverage.
