+++ src/gen_worker/clone/__init__.py
@@ -1,31 +1,13 @@
 """Public API for cloning model checkpoints from external upstreams (issue #20).
 
-# Six-method surface
-
-## High-level (one-call — 99% of tenants use these):
+Two entry points:
 
 - `clone.from_huggingface(ctx, payload)` — full pipeline (download, convert,
   upload) for a HuggingFace repo. Returns the ingested `ConversionOutput`.
 - `clone.from_civitai(ctx, payload)` — same for Civitai.
 
-## Intermediate (escape hatches for tenants building custom flows):
-
-- `clone.fetch_huggingface_snapshot(source_ref, revision, dest_dir, ...)` —
-  download-only; no upload. Returns local snapshot path.
-- `clone.fetch_civitai_file(model_version_id, file_id, dest_dir)` — download
-  a specific Civitai file.
-- `clone.parse_huggingface_metadata(source_ref)` — fetch + parse repo
-  metadata without downloading weights.
-- `clone.parse_civitai_metadata(model_version_id)` — same for Civitai.
-
-## Versioning commitment
-
-- High-level APIs (from_huggingface, from_civitai): STABLE. Signature
-  changes = major version bump.
-- Intermediate APIs: stable, narrower use-cases — changes flagged in
-  release notes; minor bump additive, major bump breaking.
-- Everything else inside `gen_worker.clone` (including `pipeline.py`,
-  `_shared.py`): LIBRARY-INTERNAL. No stability promise; refactor freely.
+Everything else inside `gen_worker.clone` (including `pipeline.py`,
+`_shared.py`) is LIBRARY-INTERNAL. No stability promise; refactor freely.
 
 ## Architectural principle
 
@@ -126,61 +108,9 @@ def from_civitai(ctx: Any, payload: Any, *, civitai_api_key: str | None = None) 
     )
 
 
-def fetch_huggingface_snapshot(
-    source_ref: str,
-    revision: str | None = None,
-    dest_dir: str | None = None,
-    *,
-    include: list[str] | None = None,
-    exclude: list[str] | None = None,
-) -> str:
-    """Download a HuggingFace snapshot to a local directory.
-
-    Intermediate API: download only; no upload, no lineage. Returns the
-    local snapshot path. For a tenant who needs "download then do something
-    custom," compose this with `ctx.save_file(...)` manually.
-    """
-    # Thin wrapper over the library's HF loader; implementation delegates to
-    # the pipeline's internal fetch path.
-    from .pipeline import _fetch_hf_snapshot_only  # type: ignore[attr-defined]
-    return _fetch_hf_snapshot_only(
-        source_ref=source_ref,
-        revision=revision,
-        dest_dir=dest_dir,
-        include=include,
-        exclude=exclude,
-    )
-
-
-def fetch_civitai_file(model_version_id: int, file_id: int, dest_dir: str) -> str:
-    """Download a specific Civitai file to dest_dir. Returns local path."""
-    from .pipeline import _fetch_civitai_file_only  # type: ignore[attr-defined]
-    return _fetch_civitai_file_only(
-        model_version_id=model_version_id,
-        file_id=file_id,
-        dest_dir=dest_dir,
-    )
-
-
-def parse_huggingface_metadata(source_ref: str, revision: str | None = None) -> dict[str, Any]:
-    """Fetch + parse HF repo metadata (library, architectures, etc.) without downloading weights."""
-    from .pipeline import _parse_hf_metadata_only  # type: ignore[attr-defined]
-    return _parse_hf_metadata_only(source_ref=source_ref, revision=revision)
-
-
-def parse_civitai_metadata(model_version_id: int) -> dict[str, Any]:
-    """Fetch + parse Civitai model-version metadata without downloading."""
-    from .pipeline import _parse_civitai_metadata_only  # type: ignore[attr-defined]
-    return _parse_civitai_metadata_only(model_version_id=model_version_id)
-
-
 __all__ = [
     "from_huggingface",
     "from_civitai",
-    "fetch_huggingface_snapshot",
-    "fetch_civitai_file",
-    "parse_huggingface_metadata",
-    "parse_civitai_metadata",
     "compute_size_facts",
     "CheckpointRef",
     "HFMeta",
